[Test] Run test against multiple gemfiles
This lets us ensure that Alba works without certain gems.
The content of `gemfiles` are three gemfiles.
One is `all` which contains all Gemfiles for Alba to work fully.
The rests are `without_active_support` and `without_oj`,
which both simulates the environment without the specified gem.
The default is `all` if `BUNDLE_GEMFILE` not specified.
If `BUNDLE_GEMFILE` is set to `without_` variants, it runs tests
only for dependencies check.

Also, now we stop `require oj` at test_helper since we load oj
if we need it anyway.