[Feat] Add `Alba.enable_root_key_transformation!`
And `root` option to `transform_keys` method.

This feature is an improvement over the previous commit,
6c18e731310183dc5ca6f626431987b4cdf685f8
It transforms root key, which is a great idea but introduces
backward incompatibility.
This commit adds the flag to control this transformation behavior,
which is default to the old behavior (not transforming root key).
See:
https://github.com/okuramasafumi/alba/pull/120#issuecomment-831780587