[Feat] Add `Resource.root_key` and `Resource.root_key!`
They are basically aliases for `key` and `key!`.
The only difference between `key` and `root_key` is that
`root_key` doesn't assign an argument when it doesn't respond to
`to_sym`. In this case it raises `NoMethodError` instead.
This assures that key is always a symbol.

`key` and `key!` methods are deprecated and will be removed
in the next major version bump.