[Refactor] Rename CONDITION_UNMET TO REMOVE_KEY
CONDITION_UNMET was used where there's no condition.
The name was wrong, so here it's renamed to REMOVE_KEY.
It was also moved to constants.rb to be required from
conditional_attribute and resource.
It's public not in public API yet. It could be useful if it's used
to remove keys in certain conditions.