[Fix] key transformation now cascades multiple levels
The cause of the bug was the usage of `Alba.assign_resource`, that
evaluate given block and return the class object. So, when it's called,
DSLs such as `attributes` are already evaluated. When we call
`transform_keys` AFTER that call, `attributes` is assigned with empty
key transformation configuration.
So I avoided this bug by not giving a block to `assign_resource` and
call `transform_keys` and then call `class_eval` manually.