[Feat] Add `prefer_resource_method!`
This is a way to change current behavior of when there are two
methods with the same name, one on the resource and one on the
object.
Some say they prefer defining methods instead of using `attribute`
since it's easier to type methods than type DSL.
The problem is that methods on resource are not called if there's
a method with the same name on the object.
I understand this situation and I'd like to support it.