[Fix] the bug of resource name inference
This bug fix is without test, unfortunately.
This bug triggers when defining the class without namespace.
In test, defining a class dynamically is not easy.
There may be a good way to test this, but for now I skip it.
The fix itself is easy, using `then` instead of `tap` makes it
return nil properly.