[Fix] Multithread bug
This is a critical bug that can corrupt the output.

The problem was the improper usage of an instance variable
in Association class.
This association instance belongs to the resource class.
This means the state in the variable is shared and can be overridden
by other threads.
The fix was to change the instance variable to the local variable.