Implement a fast path for `#attributes_to_hash`
This uses a number of performance tricks.

Rather than to call `transform_key` for every attribute of every
instances, we do it once statically, and then use `transform_values`
to fetch the values. This remove a big overhead, but is made a bit tricky
as there are a number of global configurations that can invalidate that
cache. I made it work, but things like using `ObjectSpace.each_object`
isn't reasonable.

`Hash#transform_values` is also good because it allow Ruby to directly
allocate a Hash of the right size, and skip the need to re-hash keys.
This doesn't make a big difference on the existing benchmark because
there aren't many attributes, but I expect the gains would be more
important the more attributes you have.

Makes `@_resource_methods` a Hash, so we can check whether we need to
call the method on ourself or on `obj` in `O(1)`.

Inline the `Symbol` case for fetching attributes as it's the most common.
More cases could be inlined.

Specialize a fast path when `@_on_error` isn't set, because `rescue`
clause cause an overhead (need to call `setjmp(3)`.
So it's best not to rescue when we don't have to.

Before:

```
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
                alba    82.000 i/100ms
Calculating -------------------------------------
                alba    830.409 (± 0.8%) i/s    (1.20 ms/i) -      4.182k in   5.036433s
Calculating -------------------------------------
                alba   818.241k memsize (     0.000  retained)
                         9.807k objects (     0.000  retained)
                         6.000  strings (     0.000  retained)
```

After:

```
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
                alba   107.000 i/100ms
Calculating -------------------------------------
                alba      1.074k (± 0.7%) i/s  (930.85 μs/i) -      5.457k in   5.079952s
Calculating -------------------------------------
                alba   818.241k memsize (     0.000  retained)
                         9.807k objects (     0.000  retained)
                         6.000  strings (     0.000  retained)
```