[Doc] Replace qlty badge with Codacy badge
I don't use qlty anymore since its score is too bad.
Honestly, the current code base is quite good, and CodeClimate,
the previous version of qlty, declared the score was "A".
After moving to qlty it says the score is "B".
Especially `resource` file score is "F" but all the problem is
just about size, not complexity.
qlty doesn't understand `.codeclimate.yml`, that's a deal breaker.